Move SheetHorizontalScrollContainer to own file.
