+++ src/app/dim-ui/SheetHorizontalScrollContainer.m.scss
@@ -1,0 +1,12 @@
+.horizontalScrollContainer {
+  flex: 1;
+  white-space: nowrap;
+  overflow-x: scroll;
+  overflow-y: hidden;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  -webkit-overflow-scrolling: touch;
+  contain: content;
+  touch-action: none;
+}
