+++ src/app/dim-ui/SheetHorizontalScrollContainer.m.scss.d.ts
@@ -1,8 +1,7 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'disabledScreen': string;
-  'sheetDisabled': string;
+  'horizontalScrollContainer': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
